hotwater: Live Reload for Plumber APIs

Enhances development for plumber APIs by enabling live reloading. Monitors API files for changes and automatically refreshes the server and connected web clients, allowing for faster API iteration.

See also

Author

Maintainer: Elian Thiele-Evans elianhte@gmail.com (ORCID)